refactor move models to be under api
