+++ workflow.txt
@@ -7,6 +7,6 @@
  Me: :-)
     design the various componenets
     connect the components
-    integrate several bots into a best solution
+    integrate several AI bots' respone into a single readable and efficient solution
     refactor the code
     ensure everything works
